added commander, restructure files ..
