+++ package.json
@@ -13,7 +13,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "commander": "^2.11.0",
     "foo": "^1.0.0",
+    "hankey": "^0.0.3",
     "jest": "^21.2.1",
     "node-pager": "^0.3.1",
     "shelljs": "^0.7.8"
