+++ src/tasks/diff.js
@@ -1,0 +1,6 @@
+function diff() {
+  const { stdout } = gitCmd("diff  --color=always");
+  console.log(stdout);
+}
+
+module.exports = diff;
