+++ src/tasks/index.js
@@ -1,2 +1,6 @@
 const { checkout } = require("./checkout");
-module.exports = { checkout };
+const { push } = require("./push");
+const log = require("./log");
+const diff = require("./diff");
+
+module.exports = { checkout, push, log, diff };
