+++ src/tasks/log.js
@@ -1,0 +1,8 @@
+const pager = require("node-pager");
+
+async function log() {
+  const { stdout } = gitCmd("log  --color=always");
+  pager(stdout);
+}
+
+module.exports = log;
